Added structure for test classes
