+++ src/Bruhunter.Tests/IntegrationTests/FeatureFixtureBase.cs
@@ -1,15 +1,16 @@
 ﻿using Bruhunter.Application;
 using Bruhunter.DataAccessLayer;
+using Bruhunter.Tests.Common;
 using LightBDD.XUnit2;
 using LiteDB;
 using System.IO;
 
-namespace Bruhunter.Tests
+namespace Bruhunter.Tests.IntegrationTests
 {
     public class FeatureFixtureBase : FeatureFixture
     {
         public DomainEntitiesBuilder GiveMe { get; }
-        
+
         public FeatureFixtureBase()
         {
             GiveMe = new DomainEntitiesBuilder();
