+++ src/Bruhunter.Tests/IntegrationTests/Features/AddCandidate_feature_definitions.cs
@@ -3,9 +3,9 @@ using Bruhunter.Shared.Documents;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Bruhunter.Tests
+namespace Bruhunter.Tests.IntegrationTests.Feautures
 {
-    public partial class GetCandidates_feature
+    public partial class AddCandidate_feature
     {
         private CandidateDocument candidateBeforeInsert;
 
@@ -13,15 +13,16 @@ namespace Bruhunter.Tests
         {
             candidateBeforeInsert = candidateDocument;
         }
+
         public async Task When_add_candidate()
         {
             await CandidatesService.AddCandidate(candidateBeforeInsert);
         }
 
-        public async Task Then_we_should_have_candidates()
+        public async Task Then_database_should_contain_candidate()
         {
-            var candidates = await CandidatesService.GetAllCandidates();
-            Assert.NotEmpty(candidates);
+            var candidates = await CandidatesRepository.GetAllCandidates();
+            Assert.Single(candidates);
         }
     }
 }
